requirements.txt
fastapi
uvicorn
python-multipart
pillow
python-dotenv
google-genai
pymupdf
pandas
pdfplumber
